pyproject.toml
[tool.poetry]
name = "repl_python3_GiruBot"
version = "0.1.0"
description = ""
authors = ["towzeur <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
youtube_dl = "^2021.6.6"
discord = "^1.7.3"
python-dotenv = "^0.19.2"
termcolor = "^1.1.0"
Flask = "^2.0.2"
PyNaCl = "^1.4.0"
yt-dlp = "^2021.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
